requirements.txt
flask
flask-restful
pymysql
flask-sqlalchemy
flask-migrate
flask-script
flask-jwt-extended
requests
pytest
flask-cors